make cci_gen_unique_name inline
Due to remaining circular dependencies between the API lib
and the (default) broker implementation, the core library
function cci_gen_unique_name is inline for now.

A "proper" solution would be to have a separate core library,
or to combine the broker and the API library into a single
lib by default (the main CCI implementation needs to provide
a broker anyhow).