cci_lib.vsprops: update MSVC property sheet
The MSVC property sheet for the CCI library projects is updated to
follow the other SystemC library projects more closely (e.g. TLM, SCV).

A new UserMacro has been added, intended to carry the MSVC architecture
alias used by the SystemC environment:

 *  8.0/2005 - "msvc80" (default)
 *  9.0/2008 - "msvc90"
 * 10.0/2010 - "msvc10"
 * 11.0/2012 - "msvc11"
 * 12.0/2013 - "msvc12"

The other UserMacros are renamed to omit the `_HOME` suffix, as these
names are usually intended for environment variables, pointing to the
root directory of the respective package.  The final macros here require
the correct location of the MSVC sub-directory.

 * SYSTEMC - location of SystemC MSVC directory
   default: "$(SYSTEMC_HOME)\$(MSVC)"
 * BOOST - location of Boost installation
   default: "$(BOOST_HOME)"

Rename property sheet to `cci_lib.vsprops`, as it is meant to be used in
all configurations now.