Add exception safety to parameter callback handling
In case an application-provided parameter callback throws an
exception (either via SC_REPORT_ERROR or through other means),
the nested callback protection in the callback handling of the
parameter implementation has not been properly reset.

As a result, no future callbacks of the affected callback type
are ever invoked, even if the exception is properly caught by
the application.

Switch to an RAII helper to safely clear the "oncall" tag
on all code paths (including exceptions).

Reported-by: Tom Goodfellow @thomasgoodfellow