Rename cci::detail namespace to cci::cci_impl
Most implementation-defined classes are currently placed in a
nested cci::detail namespace.  As discussed in #166, this patch
renames the namespace to cci_impl to better document the intent
and to reduce risk of naming conflict in the presence of using
directives.

Additionally, add some more doxygen suppressions.

No functional changes.