Fix inability to specify a single test(s) reported by some users
This seemed to happen only for some users but was repeatable for them.
The rewritten code avoids use of a technique that is reportedly
suscetible to "bizarre" behavior
(http://perldoc.perl.org/functions/grep.html).

Without this fix users had to either set the SYSTEMC_TEST environment
variable to the examples/ directory, or prepend the directory to
specified tests (e.g. examples/ex01_Simple_Int_Param).