rapidjson: Suppress warning on GCC 8
Compiling RapidJSON on GCC 8 gives a new compiler warning.
We treat warnings as errors in CCI, causing the CCI build
to fail on this compiler.

See Tencent/rapidjson#1246.

This is fixed in the upstream master branch of RapidJSON.
We may want to upgrade to a newer RapidJSON version in the
future.

In CCI, we can work around this by suppressing the warning
for now.