CI fixes (#325)
* ci: update ci runners

Replace deprecated runners with newer ones.
windows-2019 will be deprecated on 2025-06-30.

Signed-off-by: Antonios Salios <antonios.salios@rwth-aachen.de>

* ci: windows: install mingw-clang

Signed-off-by: Antonios Salios <antonios.salios@rwth-aachen.de>

* ci: cmake: change build type to release

Signed-off-by: Antonios Salios <antonios.salios@rwth-aachen.de>

* ci: autotools: update upload-artifact action

Signed-off-by: Antonios Salios <antonios.salios@rwth-aachen.de>

* examples(ex02): fix data type

Signed-off-by: Antonios Salios <antonios.salios@rwth-aachen.de>

* ci: only test sysc 3.0.0 for gcc/clang on windows/macos

A preprocessor directive, introdued in SystemC 2.3.4, is used in QuickThreads
asm files, but the preprocessor is not called by the build system. This
was fixed with SystemC 3.0.0.

Since this directive only make sense on Linux systems, the SystemC
build will fail on Windows/macOS for 2.3.4.

See: https://github.com/accellera-official/systemc/blob/5fc1469339775b54b1fcc2020ac744a58be5b50e/src/sysc/packages/qt/md/iX86_64.s#L76

Signed-off-by: Antonios Salios <antonios.salios@rwth-aachen.de>

* ci: disable tests for mingw gcc

The tests are segfaulting in the CI when built with MinGW GCC. They pass
when "ctest" is called directly, or when the CMake "check" target is
called in an SSH session on the CI runner...

Disable testing for this platform for now.

Signed-off-by: Antonios Salios <antonios.salios@rwth-aachen.de>

* increase test verbosity

Signed-off-by: Antonios Salios <antonios.salios@rwth-aachen.de>

---------

Signed-off-by: Antonios Salios <antonios.salios@rwth-aachen.de>